% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lau_data}
\alias{lau_data}
\alias{lau_at}
\alias{lau_be}
\alias{lau_bg}
\alias{lau_ch}
\alias{lau_cy}
\alias{lau_cz}
\alias{lau_de}
\alias{lau_dk}
\alias{lau_ee}
\alias{lau_el}
\alias{lau_es}
\alias{lau_fi}
\alias{lau_fr}
\alias{lau_hr}
\alias{lau_hu}
\alias{lau_ie}
\alias{lau_it}
\alias{lau_li}
\alias{lau_lt}
\alias{lau_lu}
\alias{lau_lv}
\alias{lau_mk}
\alias{lau_mt}
\alias{lau_nl}
\alias{lau_no}
\alias{lau_pl}
\alias{lau_pt}
\alias{lau_ro}
\alias{lau_se}
\alias{lau_si}
\alias{lau_sk}
\alias{lau_tr}
\title{Local Administrative Units (LAU) Crosswalks}
\format{
Data frames with varying columns depending on the country, typically including:
\describe{
  \item{lau_id}{Local Administrative Unit code}
  \item{lau_name}{Name of the Local Administrative Unit}
  \item{nuts_3_id}{NUTS 3 region code}
  \item{population}{Population (if available)}
}

An object of class \code{data.frame} with 2093 rows and 5 columns.

An object of class \code{data.frame} with 571 rows and 5 columns.

An object of class \code{data.frame} with 265 rows and 5 columns.

An object of class \code{data.frame} with 2135 rows and 5 columns.

An object of class \code{data.frame} with 617 rows and 5 columns.

An object of class \code{data.frame} with 6258 rows and 5 columns.

An object of class \code{data.frame} with 10972 rows and 5 columns.

An object of class \code{data.frame} with 99 rows and 5 columns.

An object of class \code{data.frame} with 79 rows and 5 columns.

An object of class \code{data.frame} with 6142 rows and 5 columns.

An object of class \code{data.frame} with 8132 rows and 5 columns.

An object of class \code{data.frame} with 309 rows and 5 columns.

An object of class \code{data.frame} with 32774 rows and 5 columns.

An object of class \code{data.frame} with 556 rows and 5 columns.

An object of class \code{data.frame} with 3155 rows and 5 columns.

An object of class \code{data.frame} with 166 rows and 5 columns.

An object of class \code{data.frame} with 7900 rows and 5 columns.

An object of class \code{data.frame} with 11 rows and 5 columns.

An object of class \code{data.frame} with 60 rows and 5 columns.

An object of class \code{data.frame} with 100 rows and 5 columns.

An object of class \code{data.frame} with 43 rows and 5 columns.

An object of class \code{data.frame} with 80 rows and 5 columns.

An object of class \code{data.frame} with 68 rows and 5 columns.

An object of class \code{data.frame} with 342 rows and 5 columns.

An object of class \code{data.frame} with 378 rows and 5 columns.

An object of class \code{data.frame} with 2477 rows and 5 columns.

An object of class \code{data.frame} with 3092 rows and 5 columns.

An object of class \code{data.frame} with 3181 rows and 5 columns.

An object of class \code{data.frame} with 290 rows and 5 columns.

An object of class \code{data.frame} with 211 rows and 5 columns.

An object of class \code{data.frame} with 2927 rows and 5 columns.

An object of class \code{data.frame} with 972 rows and 5 columns.
}
\source{
Eurostat and national statistical institutes.
}
\usage{
lau_at

lau_be

lau_bg

lau_ch

lau_cy

lau_cz

lau_de

lau_dk

lau_ee

lau_el

lau_es

lau_fi

lau_fr

lau_hr

lau_hu

lau_ie

lau_it

lau_li

lau_lt

lau_lu

lau_lv

lau_mk

lau_mt

lau_nl

lau_no

lau_pl

lau_pt

lau_ro

lau_se

lau_si

lau_sk

lau_tr
}
\description{
Datasets containing mappings from city names to LAU codes and NUTS 3 regions for various countries.
The data handles string normalization and matches cities to their respective statistical regions.
}
\keyword{datasets}
