% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nuts_data}
\alias{nuts_data}
\alias{nuts_at}
\alias{nuts_be}
\alias{nuts_bg}
\alias{nuts_ch}
\alias{nuts_cy}
\alias{nuts_cz}
\alias{nuts_de}
\alias{nuts_dk}
\alias{nuts_ee}
\alias{nuts_el}
\alias{nuts_es}
\alias{nuts_fi}
\alias{nuts_fr}
\alias{nuts_hr}
\alias{nuts_hu}
\alias{nuts_ie}
\alias{nuts_it}
\alias{nuts_li}
\alias{nuts_lt}
\alias{nuts_lu}
\alias{nuts_lv}
\alias{nuts_mk}
\alias{nuts_mt}
\alias{nuts_nl}
\alias{nuts_no}
\alias{nuts_pl}
\alias{nuts_pt}
\alias{nuts_ro}
\alias{nuts_se}
\alias{nuts_si}
\alias{nuts_sk}
\alias{nuts_tr}
\title{NUTS 3 Region Metadata}
\format{
Data frames with columns:
\describe{
  \item{nuts_3_id}{NUTS 3 region code}
  \item{nuts_3_name}{Name of the NUTS 3 region}
}

An object of class \code{data.frame} with 35 rows and 4 columns.

An object of class \code{data.frame} with 43 rows and 4 columns.

An object of class \code{data.frame} with 28 rows and 4 columns.

An object of class \code{data.frame} with 26 rows and 4 columns.

An object of class \code{data.frame} with 1 rows and 4 columns.

An object of class \code{data.frame} with 14 rows and 4 columns.

An object of class \code{data.frame} with 401 rows and 4 columns.

An object of class \code{data.frame} with 11 rows and 4 columns.

An object of class \code{data.frame} with 5 rows and 4 columns.

An object of class \code{data.frame} with 53 rows and 4 columns.

An object of class \code{data.frame} with 59 rows and 4 columns.

An object of class \code{data.frame} with 19 rows and 4 columns.

An object of class \code{data.frame} with 96 rows and 4 columns.

An object of class \code{data.frame} with 21 rows and 4 columns.

An object of class \code{data.frame} with 20 rows and 4 columns.

An object of class \code{data.frame} with 8 rows and 4 columns.

An object of class \code{data.frame} with 107 rows and 4 columns.

An object of class \code{data.frame} with 1 rows and 4 columns.

An object of class \code{data.frame} with 10 rows and 4 columns.

An object of class \code{data.frame} with 1 rows and 4 columns.

An object of class \code{data.frame} with 5 rows and 4 columns.

An object of class \code{data.frame} with 8 rows and 4 columns.

An object of class \code{data.frame} with 2 rows and 4 columns.

An object of class \code{data.frame} with 40 rows and 4 columns.

An object of class \code{data.frame} with 17 rows and 4 columns.

An object of class \code{data.frame} with 73 rows and 4 columns.

An object of class \code{data.frame} with 26 rows and 4 columns.

An object of class \code{data.frame} with 42 rows and 4 columns.

An object of class \code{data.frame} with 21 rows and 4 columns.

An object of class \code{data.frame} with 12 rows and 4 columns.

An object of class \code{data.frame} with 8 rows and 4 columns.

An object of class \code{data.frame} with 81 rows and 4 columns.
}
\source{
Eurostat
}
\usage{
nuts_at

nuts_be

nuts_bg

nuts_ch

nuts_cy

nuts_cz

nuts_de

nuts_dk

nuts_ee

nuts_el

nuts_es

nuts_fi

nuts_fr

nuts_hr

nuts_hu

nuts_ie

nuts_it

nuts_li

nuts_lt

nuts_lu

nuts_lv

nuts_mk

nuts_mt

nuts_nl

nuts_no

nuts_pl

nuts_pt

nuts_ro

nuts_se

nuts_si

nuts_sk

nuts_tr
}
\description{
Metadata for NUTS 3 regions for various countries, used for hierarchical matching.
}
\keyword{datasets}
